+++ libs/types.ts
@@ -1,19 +1,19 @@
 // TODO Replace this with a declaration file?
 
-export interface SongRequest {
-  _id: number;
-  youtubeId: string;
-  requester: string;
-  songTitle: string;
-  season: number;
-  playDate: string;
-  sotnContender: boolean;
-  sotnWinner: boolean;
-  sotsWinner: boolean;
-  artist: string;
-  featuredArtist: string;
-  songYear: string;
-}
+// export interface SongRequest {
+//   _id: number;
+//   youtubeId: string;
+//   requester: string;
+//   songTitle: string;
+//   season: number;
+//   playDate: string;
+//   sotnContender: boolean;
+//   sotnWinner: boolean;
+//   sotsWinner: boolean;
+//   artist: string;
+//   featuredArtist: string;
+//   songYear: string;
+// }
 
 export interface SongRequestDataRow {
   requester: string;
@@ -23,4 +23,14 @@ export interface SongRequestDataRow {
   sotnContender: boolean;
   sotnWinner: boolean;
   sotsWinner: boolean;
+}
+
+export interface SongRequest {
+  youtubeId: string;
+  title: string;
+  length: number;
+  requestedBy: string;
+  isBumped?: boolean;
+  isShuffled?: boolean;
+  isShuffleEntered?: boolean;
 }
